Improve API consistency and firmware compatibility
- Add #[serde(default)] to ApConfig fields that may be absent on older
  firmware (sleeptime, ble, discovery, showtimestamp, repo, env)
- Unify SaveTagConfig.invert to Option<bool> matching UploadImageOptions
- Rename TagRecord fields to idiomatic snake_case (last_seen, next_update,
  next_checkin, mode_config_json, is_external, ap_ip, update_count,
  update_last) with serde renames preserving wire compatibility

+++ src/tags.rs
@@ -69,7 +69,7 @@ impl Client {
             params.push(("lut".to_string(), lut.to_string()));
         }
         if let Some(invert) = config.invert {
-            params.push(("invert".to_string(), invert.to_string()));
+            params.push(("invert".to_string(), u8::from(invert).to_string()));
         }
 
         let body = self
